Small refactoring in c8y_configuration_plugin
