+++ plugins/c8y_configuration_plugin/src/main.rs
@@ -2,25 +2,17 @@ mod config;
 mod download;
 mod error;
 mod smartrest;
+mod upload;
 
 use crate::config::PluginConfig;
 use crate::download::handle_config_download_request;
+use crate::upload::handle_config_upload_request;
 use anyhow::Result;
 use c8y_api::http_proxy::{C8YHttpProxy, JwtAuthHttpProxy};
 use c8y_smartrest::smartrest_deserializer::SmartRestConfigDownloadRequest;
-use c8y_smartrest::{
-    smartrest_deserializer::SmartRestConfigUploadRequest,
-    smartrest_serializer::{
-        CumulocitySupportedOperations, SmartRestSerializer, SmartRestSetOperationToExecuting,
-        SmartRestSetOperationToFailed, SmartRestSetOperationToSuccessful,
-    },
-    topic::C8yTopic,
-};
-use mqtt_channel::{Connection, Message, SinkExt, StreamExt};
-use std::{
-    fs::read_to_string,
-    path::{Path, PathBuf},
-};
+use c8y_smartrest::{smartrest_deserializer::SmartRestConfigUploadRequest, topic::C8yTopic};
+use mqtt_channel::{SinkExt, StreamExt};
+use std::path::PathBuf;
 use tedge_config::{get_tedge_config, ConfigSettingAccessor, MqttPortSetting};
 use tracing::{debug, error};
 
@@ -52,91 +44,6 @@ pub async fn create_http_client() -> Result<JwtAuthHttpProxy, anyhow::Error> {
     Ok(http_proxy)
 }
 
-/// returns a c8y message specifying to set the upload config file operation status to executing.
-///
-/// example message: '501,c8y_UploadConfigFile'
-pub fn get_upload_config_file_executing_message() -> Result<Message, anyhow::Error> {
-    let topic = C8yTopic::SmartRestResponse.to_topic()?;
-    let smartrest_set_operation_status =
-        SmartRestSetOperationToExecuting::new(CumulocitySupportedOperations::C8yUploadConfigFile)
-            .to_smartrest()?;
-    Ok(Message::new(&topic, smartrest_set_operation_status))
-}
-
-/// returns a c8y SmartREST message indicating the success of the upload config file operation.
-///
-/// example message: '503,c8y_UploadConfigFile,https://{c8y.url}/etc...'
-pub fn get_upload_config_file_successful_message(
-    binary_upload_event_url: &str,
-) -> Result<Message, anyhow::Error> {
-    let topic = C8yTopic::SmartRestResponse.to_topic()?;
-    let smartrest_set_operation_status =
-        SmartRestSetOperationToSuccessful::new(CumulocitySupportedOperations::C8yUploadConfigFile)
-            .with_response_parameter(binary_upload_event_url)
-            .to_smartrest()?;
-
-    Ok(Message::new(&topic, smartrest_set_operation_status))
-}
-
-/// returns a c8y SmartREST message indicating the failure of the upload config file operation.
-///
-/// example message: '503,c8y_UploadConfigFile,https://{c8y.url}/etc...'
-pub fn get_upload_config_file_failure_message(
-    failure_reason: String,
-) -> Result<Message, anyhow::Error> {
-    let topic = C8yTopic::SmartRestResponse.to_topic()?;
-    let smartrest_set_operation_status = SmartRestSetOperationToFailed::new(
-        CumulocitySupportedOperations::C8yUploadConfigFile,
-        failure_reason,
-    )
-    .to_smartrest()?;
-
-    Ok(Message::new(&topic, smartrest_set_operation_status))
-}
-
-async fn handle_config_upload_request(
-    config_upload_request: SmartRestConfigUploadRequest,
-    mqtt_client: &mut Connection,
-    http_client: &mut JwtAuthHttpProxy,
-) -> Result<()> {
-    // set config upload request to executing
-    let msg = get_upload_config_file_executing_message()?;
-    let () = mqtt_client.published.send(msg).await?;
-
-    let upload_result = upload_config_file(
-        Path::new(config_upload_request.config_type.as_str()),
-        http_client,
-    )
-    .await;
-    match upload_result {
-        Ok(upload_event_url) => {
-            let successful_message = get_upload_config_file_successful_message(&upload_event_url)?;
-            let () = mqtt_client.published.send(successful_message).await?;
-        }
-        Err(err) => {
-            let failed_message = get_upload_config_file_failure_message(err.to_string())?;
-            let () = mqtt_client.published.send(failed_message).await?;
-        }
-    }
-
-    Ok(())
-}
-
-async fn upload_config_file(
-    config_file_path: &Path,
-    http_client: &mut JwtAuthHttpProxy,
-) -> Result<String> {
-    // read the config file contents
-    let config_content = read_to_string(config_file_path)?;
-
-    // upload config file
-    let upload_event_url = http_client
-        .upload_config_file(config_file_path, &config_content)
-        .await?;
-
-    Ok(upload_event_url)
-}
-
 #[tokio::main]
 async fn main() -> Result<(), anyhow::Error> {
     tedge_utils::logging::initialise_tracing_subscriber(LOG_LEVEL_DEBUG);
@@ -157,9 +64,9 @@ async fn main() -> Result<(), anyhow::Error> {
         if let Ok(payload) = message.payload_str() {
             let result = match payload.split(',').next().unwrap_or_default() {
                 "524" => {
-                    debug!("{}", payload);
                     let config_download_request =
                         SmartRestConfigDownloadRequest::from_smartrest(payload)?;
+
                     handle_config_download_request(
                         &plugin_config,
                         config_download_request,
@@ -169,7 +76,6 @@ async fn main() -> Result<(), anyhow::Error> {
                     .await
                 }
                 "526" => {
-                    debug!("{}", payload);
                     // retrieve config file upload smartrest request from payload
                     let config_upload_request =
                         SmartRestConfigUploadRequest::from_smartrest(payload)?;
