+++ plugins/c8y_configuration_plugin/src/smartrest.rs
@@ -1,26 +1,7 @@
-use crate::config::PluginConfig;
 use c8y_smartrest::error::SmartRestSerializerError;
 use c8y_smartrest::smartrest_serializer::SmartRest;
 use c8y_smartrest::topic::C8yTopic;
 use mqtt_channel::Message;
-
-impl PluginConfig {
-    pub fn to_message(&self) -> Result<Message, anyhow::Error> {
-        let topic = C8yTopic::SmartRestResponse.to_topic()?;
-        Ok(Message::new(&topic, self.to_smartrest_payload()))
-    }
-
-    // 119,typeA,typeB,...
-    fn to_smartrest_payload(&self) -> String {
-        let config_types = self
-            .files
-            .iter()
-            .map(|x| x.to_string())
-            .collect::<Vec<_>>()
-            .join(",");
-        format!("119,{config_types}")
-    }
-}
 
 pub trait GetSmartRestMessage {
     fn executing() -> Result<Message, SmartRestSerializerError> {
@@ -28,8 +9,8 @@ pub trait GetSmartRestMessage {
         Ok(Self::create_message(status))
     }
 
-    fn successful() -> Result<Message, SmartRestSerializerError> {
-        let status = Self::status_successful()?;
+    fn successful(parameter: Option<String>) -> Result<Message, SmartRestSerializerError> {
+        let status = Self::status_successful(parameter)?;
         Ok(Self::create_message(status))
     }
 
@@ -44,31 +25,6 @@ pub trait GetSmartRestMessage {
     }
 
     fn status_executing() -> Result<SmartRest, SmartRestSerializerError>;
-    fn status_successful() -> Result<SmartRest, SmartRestSerializerError>;
+    fn status_successful(parameter: Option<String>) -> Result<SmartRest, SmartRestSerializerError>;
     fn status_failed(failure_reason: String) -> Result<SmartRest, SmartRestSerializerError>;
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use test_case::test_case;
-
-    #[test_case(
-        PluginConfig {
-            files: vec!["typeA".to_string()]
-        },
-        "119,typeA".to_string()
-        ;"single file"
-    )]
-    #[test_case(
-        PluginConfig {
-        files: vec!["typeA".to_string(), "typeB".to_string(), "typeC".to_string()]
-        },
-        "119,typeA,typeB,typeC".to_string()
-        ;"multiple files"
-    )]
-    fn get_smartrest(input: PluginConfig, expected_output: String) {
-        let output = input.to_smartrest_payload();
-        assert_eq!(output, expected_output);
-    }
 }
